refactor: db container per integration test file
